+++ CHANGELOG.md
@@ -6,11 +6,55 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
 ## [Unreleased]
 
-## [0.20200106.1] - 2020-10-20
+## [0.20210401] - 2021-04-01
+
+### Fixed
+- chroot: fix tedious shell glitches in clara chroot using real shell
+
+## [0.20210305] - 2021-03-05
+
+### Fixed
+- ipmi: make sure the escape sequence is the same whether or not a jump host is used
+- ipmi: use localhost as the conman destination when using a jump host
+- ipmi: fix a missing import
+
+## [0.20210224] - 2021-02-24
+
+### Added
+- ipmi: Option to use an ssh connexion as a relay to the conman server, set to true by default
+- images: ensure user's read access to /var/lib/rpm on rhel-based generated images
+
+## [0.20201203] - 2020-12-03
 
 ### Fixed
 
+- images: Improve Yum/DNF support
 - repo: control digest used to decode repos signature key
+
+## [0.20201015] - 2020-10-15
+
+### Fixed
+- clara: fix symlink run and var/run
+- clara: fix input function for python3
+
+## [0.20201007] - 2020-10-07
+
+### Added
+- images: Add list_repos feature for rhel distrib to support multiple repos
+- images: symlink run -> var/run
+
+## [0.20200707] - 2020-07-07
+
+### Added
+- images: add support for centos8 image creation
+
+## [0.20200617] - 2020-06-17
+
+### Added
+- clara: add pytest unitary tests
+
+### Changed
+- clara: port from Python2 to Python3.
 
 ## [0.20200106] - 2020-01-06